Fix new router example still using old router for inner child routes.
Split router config into different files.
