+++ example-apps/kitchen-sink-new-router/source/containers/kitchen-sink.ts
@@ -1,55 +1,9 @@
 import { Component, Inject, OnInit, OnDestroy }
  from '@angular/core';
-import { ROUTER_DIRECTIVES, RouterConfig, Router, ActivatedRoute}
+import { ROUTER_DIRECTIVES, Router, ActivatedRoute}
  from '@angular/router';
 
 import {CamelCasePipe} from '../pipes/camelcase';
-
-import Home from '../components/home';
-import InputOutput from '../components/input-output/input-output';
-import MyForm from '../components/form-controls/my-form';
-import Form2 from '../components/form-controls/form2';
-import DynamicControls from '../components/dynamic-controls/dynamic-controls';
-import ControlForm from '../components/form-controls/control-form';
-import TodoApp from '../components/todo-app/todo-app';
-import DITree from '../components/di-tree/di-tree';
-import ChangeDetection from '../components/change-detection/change-detection';
-import AngularDirectives from
-  '../components/angular-directives/angular-directives';
-import Demo from '../components/demo/demo';
-import StressTester from '../components/stress-tester/stress-tester';
-
-import Start from '../components/router/start';
-import StartChild from '../components/router/start-child';
-import StartMain from '../components/router/start-main';
-import InnerChild from '../components/router/inner-child';
-import RouterData1 from '../components/router/router-data1';
-import RouterData2 from '../components/router/router-data2';
-import AuxComp from '../components/router/aux-comp';
-import MetadataTest from '../components/metadata-test/metadata-test';
-
-export const KitchenSinkRoutes: RouterConfig = [
-  { path: '', component: Home },
-  { path: 'input-output', component: InputOutput },
-  { path: 'my-form', component: MyForm },
-  { path: 'form2', component: Form2 },
-  { path: 'control-form', component: ControlForm },
-  { path: 'dynamic-controls', component: DynamicControls },
-  { path: 'todo-app', component: TodoApp },
-  { path: 'di-tree', component: DITree },
-  { path: 'angular-directives', component: AngularDirectives },
-  { path: 'change-detection', component: ChangeDetection },
-  { path: 'demo', component: Demo },
-  { path: 'stress-tester', component: StressTester },
-  { path: 'start', component: Start, children: [
-    { path: 'main', component: StartMain },
-    { path: 'auxcomp', component: AuxComp, outlet: 'aux' },
-    { path: 'child', component: StartChild },
-    { path: 'router-data1/:name', component: RouterData1 },
-    { path: 'router-data2/:name/:message', component: RouterData2 }]
-  },
-  { path: 'metadata-test', component: MetadataTest },
-];
 
 @Component({
   selector: 'kitchen-sink',
